refactor: pull demofs into lib crate
