+++ src/fs/mod.rs
@@ -1,0 +1,3 @@
+//! Defines filesystem implememntations
+
+mod fake;
