+++ src/tcp.rs
@@ -8,14 +8,13 @@
 //!
 //! The implementation supports configurable export paths and notification
 //! on mount/unmount operations.
+
 use std::collections::HashSet;
 use std::net::SocketAddr;
-use std::sync::atomic::AtomicU64;
 use std::sync::Arc;
 use std::time::Duration;
 use std::{io, net::IpAddr};
 
-use async_trait::async_trait;
 use dashmap::DashMap;
 use tokio::io::AsyncWriteExt;
 use tokio::net::TcpListener;
@@ -47,7 +46,7 @@ pub type NFSExportTable = DashMap<crate::xdr::nfs3::fs_id, NFSExportTableEntry>;
 /// NFS TCP Connection Handler that listens for incoming NFS client connections
 /// and processes RPC messages over TCP transport.
 pub struct NFSTcpListener {
-    next_id: AtomicU64,
+    next_id: u64,
     /// TCP Listener for accepting incoming connections
     listener: TcpListener,
     /// Port on which the server is listening
@@ -139,60 +138,6 @@ async fn process_socket(
     }
 }
 
-/// Interface for NFS TCP servers that defines common operations
-/// for managing and interacting with NFS clients over TCP connections.
-///
-/// This trait provides methods for:
-/// - Getting information about the listening socket
-/// - Setting up mount event notifications
-/// - Starting the server to process client connections
-#[async_trait]
-pub trait NFSTcp: Send + Sync {
-    /// Returns the actual port number on which the server is listening
-    ///
-    /// This is especially useful when binding to port 0, which allows the OS
-    /// to assign any available port. After binding, this method can be used
-    /// to determine which port was actually assigned.
-    fn get_listen_port(&self) -> u16;
-
-    /// Returns the IP address on which the server is listening
-    ///
-    /// This is useful when the server binds to a wildcard address (0.0.0.0 or ::)
-    /// or when using the "auto" IP address feature, to determine the actual
-    /// network interface being used.
-    fn get_listen_ip(&self) -> IpAddr;
-
-    /// Registers a channel to receive notifications about mount and unmount events
-    ///
-    /// # Arguments
-    ///
-    /// * `signal` - MPSC sender that will receive boolean values:
-    ///   * `true` when a client mounts the file system
-    ///   * `false` when a client unmounts the file system
-    ///
-    /// # Returns
-    /// `io::Result<()>` indicating:
-    /// - `Ok(())` on successful operation
-    /// - `Err` if the export ID is not found in the export table
-    async fn set_mount_listener(
-        &mut self,
-        fs_id: xdr::nfs3::fs_id,
-        signal: mpsc::Sender<bool>,
-    ) -> io::Result<()>;
-
-    /// Starts the NFS server and processes client connections
-    ///
-    /// This method:
-    /// - Accepts incoming TCP connections from NFS clients
-    /// - Creates a new RPC context for each connection
-    /// - Spawns an asynchronous task to handle each connection
-    /// - Continues accepting connections indefinitely
-    ///
-    /// This method runs in an infinite loop and only returns if there's an error
-    /// with the underlying TCP listener.
-    async fn handle_forever(&self) -> io::Result<()>;
-}
-
 impl NFSTcpListener {
     /// Creates a new NFS TCP listener bound to the specified IP address and port
     ///
@@ -246,7 +191,7 @@ impl NFSTcpListener {
         };
 
         Ok(NFSTcpListener {
-            next_id: AtomicU64::new(0),
+            next_id: 0,
             listener,
             port,
             export_table: Arc::new(DashMap::new()),
@@ -299,7 +244,7 @@ impl NFSTcpListener {
         T: NFSFileSystem + Send + Sync + 'static,
         S: AsRef<str>,
     {
-        let fs_id = self.next_id.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
+        let fs_id = self.next_id();
 
         let export_name = format!("/{}", export_name.as_ref().trim_matches('/'));
 
@@ -338,16 +283,13 @@ impl NFSTcpListener {
         }
         Ok(())
     }
-}
 
-#[async_trait]
-impl NFSTcp for NFSTcpListener {
     /// Returns the actual port number on which the server is listening
     ///
     /// This is especially useful when binding to port 0, which allows the OS
     /// to assign any available port. After binding, this method can be used
     /// to determine which port was actually assigned.
-    fn get_listen_port(&self) -> u16 {
+    pub fn get_listen_port(&self) -> u16 {
         let addr = self.listener.local_addr().unwrap();
         addr.port()
     }
@@ -357,7 +299,7 @@ impl NFSTcp for NFSTcpListener {
     /// This is useful when the server binds to a wildcard address (0.0.0.0 or ::)
     /// or when using the "auto" IP address feature, to determine the actual
     /// network interface being used.
-    fn get_listen_ip(&self) -> IpAddr {
+    pub fn get_listen_ip(&self) -> IpAddr {
         let addr = self.listener.local_addr().unwrap();
         addr.ip()
     }
@@ -374,7 +316,7 @@ impl NFSTcp for NFSTcpListener {
     /// `io::Result<()>` indicating:
     /// - `Ok(())` on successful operation
     /// - `Err` if the export ID is not found in the export table
-    async fn set_mount_listener(
+    pub async fn set_mount_listener(
         &mut self,
         fs_id: xdr::nfs3::fs_id,
         signal: mpsc::Sender<bool>,
@@ -394,7 +336,7 @@ impl NFSTcp for NFSTcpListener {
     ///
     /// This method runs in an infinite loop and only returns if there's an error
     /// with the underlying TCP listener.
-    async fn handle_forever(&self) -> io::Result<()> {
+    pub async fn handle_forever(&self) -> io::Result<()> {
         loop {
             let (socket, _) = self.listener.accept().await?;
             let context = rpc::Context {
@@ -412,5 +354,10 @@ impl NFSTcp for NFSTcpListener {
                 let _ = process_socket(socket, context).await;
             });
         }
+    }
+
+    fn next_id(&mut self) -> u64 {
+        self.next_id += 1;
+        self.next_id
     }
 }
